[converter] Commented out old conversion function which converts to tm [converter / task list] solved all the compilation warnings

+++ 4get/4get/logic_task_list.h
@@ -26,7 +26,7 @@ private:
 
 	list<Task>::iterator getIterator(list<Task>& insertionList, Task task);
 	list<Task>::iterator iterateToTask(list<Task>& listToEdit, int task);
-	void findID(list<Task>* listToEdit, int IDNumber, list<Task>::iterator& iterator);
+	void findID(list<Task>* listToEdit, long long IDNumber, list<Task>::iterator& iterator);
 	list<Task>* determineList(ListType listType);
 
 public:
@@ -39,7 +39,7 @@ public:
 	//Public list manipulation functions
 	bool addToList(Task task, ListType listToAdd);
 	void deleteFromList(int taskToDelete);
-	void deleteIDFromList(int IDNumber, ListType listToDelete);
+	void deleteIDFromList(long long IDNumber, ListType listToDelete);
 	bool markDone(int taskToMark);
 
 	//Getter functions
